feat: Update guest login #201
