+++ server/package.json
@@ -23,6 +23,7 @@
     "mongoose": "^6.8.3",
     "nodemailer": "^6.9.1",
     "passport": "^0.6.0",
+    "passport-custom": "^1.1.1",
     "passport-local": "^1.0.0",
     "redis": "^4.6.5",
     "session-file-store": "^1.5.0"
